requirements.txt
# ============================================================
#   EARTHLENS AI – REQUIREMENTS
#   Install: pip install -r requirements.txt
# ============================================================


# ------------------------------------------------------------
# 🖥️ WEB UI
# ------------------------------------------------------------
streamlit>=1.35.0


# ------------------------------------------------------------
# 🛰️ SATELLITE DATA & GEOSPATIAL
# ------------------------------------------------------------
sentinelhub>=3.10.0          # Sentinel Hub API client
# landsatxplore REMOVED      # conflicts with shapely>=2.0 — using requests + USGS M2M API directly
rasterio>=1.3.9              # Read/write raster (GeoTIFF) files
pyproj>=3.6.1                # CRS transformations
shapely>=2.0.4               # Geometric operations
geopandas>=0.14.4            # Geospatial DataFrames


# ------------------------------------------------------------
# 🔢 NUMERICAL & SCIENTIFIC
# ------------------------------------------------------------
numpy>=1.26.0
scipy>=1.13.0


# ------------------------------------------------------------
# 🖼️ IMAGE PROCESSING
# ------------------------------------------------------------
Pillow>=10.3.0               # Basic image I/O
opencv-python-headless>=4.9.0.80   # Advanced image processing (headless for server)
scikit-image>=0.23.2         # Scientific image analysis


# ------------------------------------------------------------
# 📊 DATA HANDLING
# ------------------------------------------------------------
pandas>=2.2.0
xarray>=2024.3.0             # N-dimensional arrays (satellite bands)
netCDF4>=1.6.5               # NetCDF file support


# ------------------------------------------------------------
# 🤖 MACHINE LEARNING
# ------------------------------------------------------------
scikit-learn>=1.4.2          # Classical ML (classifier)
joblib>=1.4.0                # Model serialization


# ------------------------------------------------------------
# 🗺️ VISUALIZATION
# ------------------------------------------------------------
folium>=0.16.0               # Interactive maps
streamlit-folium>=0.20.0     # Folium inside Streamlit
plotly>=5.22.0               # Interactive charts
matplotlib>=3.9.0            # Static plots / colormaps
branca>=0.7.2                # Folium colormap support


# ------------------------------------------------------------
# 🌐 HTTP & APIs
# ------------------------------------------------------------
requests>=2.32.0
httpx>=0.27.0                # Async HTTP client
aiohttp>=3.9.5               # Async requests (batch fetching)


# ------------------------------------------------------------
# 🔐 ENVIRONMENT & CONFIG
# ------------------------------------------------------------
python-dotenv>=1.0.1         # Load .env files
pydantic>=2.7.0              # Settings validation (optional upgrade path)


# ------------------------------------------------------------
# 🗜️ FILE FORMATS
# ------------------------------------------------------------
h5py>=3.11.0                 # HDF5 support
zarr>=2.17.2                 # Chunked array storage


# ------------------------------------------------------------
# ⚡ CACHING & PERFORMANCE
# ------------------------------------------------------------
cachetools>=5.3.3            # In-memory caching utilities
diskcache>=5.6.3             # Disk-based cache for large imagery


# ------------------------------------------------------------
# 🧪 DEVELOPMENT & TESTING
# ------------------------------------------------------------
jupyter>=1.0.0               # earthlens_lab notebooks
ipykernel>=6.29.4
pytest>=8.2.0
pytest-cov>=5.0.0


# ------------------------------------------------------------
# 📝 LOGGING & UTILITIES
# ------------------------------------------------------------
loguru>=0.7.2                # Clean logging
tqdm>=4.66.4                 # Progress bars (scripts / notebooks)
rich>=13.7.1                 # Pretty terminal output


# ============================================================
# NOTES
# ============================================================
# Python version recommended: 3.10 – 3.12
#
# For GPU-accelerated ML (optional, Phase 4):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# GDAL (if needed for advanced raster ops):
#   conda install -c conda-forge gdal
#   OR: pip install GDAL==$(gdal-config --version)
# ============================================================
